refactor queenMachine into gameMachine
