+++ src/Game/Blobs/blobLarva/index.ts
@@ -1,2 +1,3 @@
 export * from './blobLarva';
 export * from './types';
+export * from './draw';
