+++ src/Game/Blobs/bloblet/index.ts
@@ -1,2 +1,3 @@
 export * from './bloblet';
 export * from './types';
+export * from './draw';
