+++ src/Game/Blobs/index.ts
@@ -1,3 +1,3 @@
 export * from './bloblet/bloblet';
 export * from './blobLarva/blobLarva';
-export * from './blobQueen';
+export * from '../gameMachine';
