+++ src/Game/Resources/shrub/types.ts
@@ -1,6 +1,8 @@
 import { ActorRefFrom, StateMachine } from 'xstate';
 import { Coordinates, PersistedActor, DrawEvent } from 'game/types';
 
+export type { DrawEvent };
+
 export type Context = {
   id: string;
   position: Coordinates;
