+++ src/Game/animations/index.ts
@@ -1,1 +1,2 @@
 export { makeShowNumber } from './showNumber';
+export * from './animationMachine';
