+++ src/Game/gameMachine/index.ts
@@ -1,0 +1,2 @@
+export * from './gameMachine';
+export * from './types';
