+++ src/Game/paramaters.ts
@@ -12,6 +12,8 @@ export const QUEEN_POSITION = {
   x: WORLD_WIDTH / 2,
   y: WORLD_HEIGHT / 2,
 };
+export const QUEEN_RADIUS_X = 60;
+export const QUEEN_RADIUS_Y = 40;
 
 // Larvae
 export const MAX_LARVAE = 4;
